requirements.txt
# Python 3.11+ recommended

pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

psycopg2-binary>=2.9.0

requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0

nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.4.0

scipy>=1.10.0

matplotlib>=3.7.0
seaborn>=0.12.0

wandb>=0.15.0

# Optional, for local model inference
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0